# Marks contexts extracted by AWSPropagation and cuts overhead