This marks contexts extracted by AWSPropagation so that calls to
`rootField` are reliable even when AWS isn't the current propagation
factory.

This also dramatically reduces extraction overhead by avoiding
allocation of constants, and early returning when required fields are
absent.

Before:
```
Benchmark                                   Mode  Cnt    Score   Error   Units
PropagationBenchmarks.extract_aws          thrpt   15    4.145 ± 0.292  ops/us
PropagationBenchmarks.extract_aws_nothing  thrpt   15  132.670 ± 1.465  ops/us
PropagationBenchmarks.extract_b3           thrpt   15   10.697 ± 0.394  ops/us
PropagationBenchmarks.extract_b3_nothing   thrpt   15  170.315 ± 5.142  ops/us
```

After: (higher numbers are better)
```
Benchmark                                   Mode  Cnt    Score    Error   Units
PropagationBenchmarks.extract_aws          thrpt   15    4.256 ± 0.245  ops/us
PropagationBenchmarks.extract_aws_nothing  thrpt   15  214.205 ±  4.829  ops/us
PropagationBenchmarks.extract_b3           thrpt   15   10.769 ±  0.316  ops/us
PropagationBenchmarks.extract_b3_nothing   thrpt   15  268.972 ± 12.388  ops/us
```